refactor(frontend): split PasswordSettings into focused dialog subcomponents
password-settings.tsx grew to 445 lines holding five react-hook-form
instances, four submit handlers, and four dialogs in a single component.
Touching any one auth flow (bootstrap-token setup, change, remove,
proxy-auth re-verify with optional TOTP) meant scrolling past all the
others and risked cross-form state coupling via the shared activeDialog/
verifyStep/error trio.

Extract one self-contained component per dialog and let each own its
form + error + (where relevant) sub-step state:

- password-setup-dialog.tsx  – setup form, bootstrap-token field gated
  on bootstrapRequired, bootstrap warning AlertMessage.
- password-change-dialog.tsx – current/new password form.
- password-remove-dialog.tsx – confirm-current-password remove form
  with destructive submit button.
- password-verify-dialog.tsx – two-step password→TOTP flow, owns its
  own verifyStep so the parent no longer needs to track it.

password-settings.tsx is now layout-only: reads enough auth-store flags
to pick which action button to show, drives an `activeDialog` discriminant,
and mounts all four subcomponents.

No DOM, copy, route, or store-write change. password-settings.test.tsx
(10 tests covering setup/change/remove/error/auth-mode visibility)
continues to pass unchanged.

Closes #617
